Refactor graphql folder structure
