+++ apps/skilavottord/web/graphql/mutations/GDPR.ts
