+++ apps/skilavottord/web/graphql/mutations/index.ts
@@ -1,1 +1,1 @@
-export * from './gdprInfo'
+export * from './GDPR'
