+++ apps/skilavottord/web/graphql/queries/GDPR.ts
