+++ apps/skilavottord/web/graphql/queries/User.ts
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag'
+
+export const GET_USER = gql`
+  query UserQuery {
+    user {
+      name
+      nationalId
+      mobile
+    }
+  }
+`
