+++ apps/skilavottord/web/graphql/queries/Vehicles.ts
