--- apps/skilavottord/web/graphql/queries/allRecyclingPartners.ts
@@ -1,18 +1,0 @@
-import gql from 'graphql-tag'
-
-export const GET_ALL_RECYCLING_PARTNERS = gql`
-  query getAllRecyclingPartners {
-    getAllRecyclingPartners {
-      companyId
-      companyName
-      address
-      postnumber
-      city
-      website
-      phone
-      active
-      createdAt
-      updatedAt
-    }
-  }
-`
