+++ apps/skilavottord/web/graphql/queries/index.ts
@@ -1,5 +1,4 @@
-export * from './vehicles'
-export * from './recyclingPartner'
-export * from './activeRecyclingPartners'
-export * from './allRecyclingPartners'
-export * from './gdprInfo'
+export * from './Vehicles'
+export * from './RecylingPartners'
+export * from './GDPR'
+export * from './User'
